requirements.txt
fastapi
uvicorn
python-telegram-bot
requests
python-dotenv
apscheduler
openai>=1.0.0
yfinance
ta